volume_flex_card/requirements.txt
requests
pandas
python-dotenv
web3>=6.0.0
redis